requirements.txt
transformers==4.36.2
torch
torchvision
pillow
gradio
accelerate
